[Fix] 펜스분리, 아이템삭제, 음식수치 올라가게 수정
Door가 열리지 않는 현상이 일어나 수진님 코드 수정하여 열리게 코드 임시 수정함
펜스 점프모션 조정
높은펜스 Tag 분리하여 점프모션 나오지 않게 조정

+++ Assets/3.Script/ParkJun/StatusController.cs
@@ -132,7 +132,7 @@ public class StatusController : MonoBehaviour
             
     }
     
-    public void increaseHP(int _count)
+    public void increaseHP(float _count)
     {
         if (currentHp + _count < hp)
         {
